+++ app/src/main/java/pl/polsl/laboratorioweobecnosci/database/models/Student.kt
@@ -15,5 +15,4 @@ data class Student(
     @ColumnInfo(name = "workstation_id")
     val workstationId:Int,
     @PrimaryKey(autoGenerate = true)
-    val id: Long = 0L
-)
+    val id: Int = 0
